+++ crates/cheatnet/tests/cheatcodes/mod.rs
@@ -3,5 +3,6 @@ mod mock_call;
 mod prank;
 mod precalculate_address;
 mod roll;
+mod spoof;
 mod spy_events;
 mod warp;