+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - rename `TxtParser` trait to `FileParser`
 - rename `parse_txt` trait to `read_txt`
 - support for printing in contracts
+- `spoof` cheatcode
 
 ### Cast
 
@@ -30,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for custom networks - accounts created on custom networks are saved in `accounts-file` under network's
   chain_id
 - `accounts-file` field in Scarb.toml profile
+- Include the class hash of an account contract in the `accounts-file`
 
 #### Removed